Grapple Indicator
I made the crosshair change to red if the player cannot grapple (no object in range, currently grappling or on a grapple cooldown) and change to green if the player is able to grapple.

This is temporary as in the main game the main crosshair will not be the indicator as to if the player can grapple or not - I just thought to implement the system as one piece of feedback we gathered in our User Test was that it was not clear whether something could be grappled or not.